backend/requirements.txt
Flask
python-dotenv
requests
requests-oauthlib
Flask-Cors
Flask-Session
msal

# prod
gunicorn